lib/tasks/boll/requirements.txt
tqsdk==2.4.1
tushare==1.2.62
joblib
pyyaml